fix(#5015): bound headless runtime mismatch recovery
