+++ src/services/discord/router/message_handler/provider_isolation.rs
@@ -85,521 +85,6 @@ pub(super) fn metadata_delivery_bot(metadata: Option<&serde_json::Value>) -> Opt
         .and_then(normalize_delivery_bot_name)
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(super) enum RuntimeKindEvidenceStrength {
-    Weak,
-    Moderate,
-    Strong,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(super) struct ManagedRuntimeExpectation {
-    pub(super) runtime_kind: RuntimeHandoffKind,
-    evidence_strength: RuntimeKindEvidenceStrength,
-}
-
-#[cfg(unix)]
-pub(super) fn prelaunch_runtime_kind_for_managed_session(
-    provider: &ProviderKind,
-    remote_profile_is_none: bool,
-    has_tmux_session_name: bool,
-    channel_id: Option<u64>,
-) -> Option<ManagedRuntimeExpectation> {
-    if !remote_profile_is_none
-        || !has_tmux_session_name
-        || !provider.uses_managed_tmux_backend()
-        || !claude::is_tmux_available()
-    {
-        return None;
-    }
-    let selection =
-        crate::services::provider_hosting::resolve_provider_session_selection_with_channel(
-            provider, true, channel_id,
-        );
-    let hook_endpoint_present =
-        crate::services::claude_tui::hook_server::current_hook_endpoint().is_some();
-    let runtime_kind = if selection.driver
-        == crate::services::provider_hosting::ProviderSessionDriver::TuiHosting
-    {
-        match provider {
-            ProviderKind::Claude if hook_endpoint_present => RuntimeHandoffKind::ClaudeTui,
-            ProviderKind::Codex => RuntimeHandoffKind::CodexTui,
-            _ => RuntimeHandoffKind::LegacyTmuxWrapper,
-        }
-    } else {
-        RuntimeHandoffKind::LegacyTmuxWrapper
-    };
-    let evidence_strength = if matches!(provider, ProviderKind::Claude)
-        && selection.driver == crate::services::provider_hosting::ProviderSessionDriver::TuiHosting
-        && !hook_endpoint_present
-    {
-        RuntimeKindEvidenceStrength::Weak
-    } else {
-        RuntimeKindEvidenceStrength::Strong
-    };
-    Some(ManagedRuntimeExpectation {
-        runtime_kind,
-        evidence_strength,
-    })
-}
-
-#[cfg(not(unix))]
-pub(super) fn prelaunch_runtime_kind_for_managed_session(
-    _provider: &ProviderKind,
-    _remote_profile_is_none: bool,
-    _has_tmux_session_name: bool,
-    _channel_id: Option<u64>,
-) -> Option<ManagedRuntimeExpectation> {
-    None
-}
-
-/// Seeds the durable inflight runtime fields before runtime handoff binds the
-/// provider-owned transcript path.
-fn prelaunch_inflight_runtime_seed_from_paths(
-    tmux_name: &str,
-    output_path: String,
-    input_fifo_path: String,
-    session_exists: bool,
-    prelaunch_runtime_kind: Option<RuntimeHandoffKind>,
-) -> (Option<String>, Option<String>, Option<String>, u64) {
-    let is_claude_tui = prelaunch_runtime_kind == Some(RuntimeHandoffKind::ClaudeTui);
-    let last_offset = (!is_claude_tui)
-        .then(|| {
-            std::fs::metadata(&output_path)
-                .map(|metadata| metadata.len())
-                .unwrap_or(0)
-        })
-        .unwrap_or(0);
-    (
-        Some(tmux_name.to_string()),
-        (!is_claude_tui).then_some(output_path),
-        Some(input_fifo_path),
-        session_exists.then_some(last_offset).unwrap_or(0),
-    )
-}
-
-pub(super) fn prelaunch_inflight_runtime_seed(
-    provider: &ProviderKind,
-    remote_profile_is_none: bool,
-    tmux_session_name: Option<&str>,
-    prelaunch_runtime_kind: Option<RuntimeHandoffKind>,
-) -> (Option<String>, Option<String>, Option<String>, u64) {
-    #[cfg(unix)]
-    {
-        if remote_profile_is_none
-            && provider.uses_managed_tmux_backend()
-            && claude::is_tmux_available()
-            && let Some(tmux_name) = tmux_session_name
-        {
-            let (output_path, input_fifo_path) = tmux_runtime_paths(tmux_name);
-            let session_exists =
-                crate::services::tmux_diagnostics::tmux_session_has_live_pane(tmux_name);
-            return prelaunch_inflight_runtime_seed_from_paths(
-                tmux_name,
-                output_path,
-                input_fifo_path,
-                session_exists,
-                prelaunch_runtime_kind,
-            );
-        }
-    }
-    #[cfg(not(unix))]
-    {
-        let _ = (
-            provider,
-            remote_profile_is_none,
-            tmux_session_name,
-            prelaunch_runtime_kind,
-        );
-    }
-    (None, None, None, 0)
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(super) struct ObservedManagedRuntimeKind {
-    pub(super) runtime_kind: RuntimeHandoffKind,
-    evidence_strength: RuntimeKindEvidenceStrength,
-}
-
-#[cfg(unix)]
-pub(super) fn observed_runtime_kind_for_managed_tmux(
-    provider: &ProviderKind,
-    tmux_session_name: &str,
-) -> Option<ObservedManagedRuntimeKind> {
-    if let Some(binding) =
-        crate::services::tui_prompt_dedupe::runtime_binding_for_tmux_session(tmux_session_name)
-    {
-        return Some(ObservedManagedRuntimeKind {
-            runtime_kind: binding.runtime_kind,
-            evidence_strength: RuntimeKindEvidenceStrength::Strong,
-        });
-    }
-    if let Some(marker) =
-        crate::services::tmux_common::resolve_tmux_runtime_kind_marker(tmux_session_name)
-    {
-        return Some(ObservedManagedRuntimeKind {
-            runtime_kind: marker,
-            evidence_strength: RuntimeKindEvidenceStrength::Strong,
-        });
-    }
-    if crate::services::tmux_common::resolve_session_temp_path(tmux_session_name, "input").is_some()
-    {
-        return Some(ObservedManagedRuntimeKind {
-            runtime_kind: RuntimeHandoffKind::LegacyTmuxWrapper,
-            evidence_strength: RuntimeKindEvidenceStrength::Moderate,
-        });
-    }
-    match provider {
-        ProviderKind::Claude => Some(ObservedManagedRuntimeKind {
-            runtime_kind: RuntimeHandoffKind::ClaudeTui,
-            evidence_strength: RuntimeKindEvidenceStrength::Weak,
-        }),
-        ProviderKind::Codex => Some(ObservedManagedRuntimeKind {
-            runtime_kind: RuntimeHandoffKind::CodexTui,
-            evidence_strength: RuntimeKindEvidenceStrength::Weak,
-        }),
-        _ => None,
-    }
-}
-
-#[cfg(unix)]
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(super) struct LiveTuiProviderSessionRecovery {
-    session_id: String,
-    output_path: String,
-}
-
-#[cfg(unix)]
-pub(super) fn live_tui_provider_session_recovery(
-    provider: &ProviderKind,
-    tmux_session_name: Option<&str>,
-) -> Option<LiveTuiProviderSessionRecovery> {
-    if !matches!(provider, ProviderKind::Claude) {
-        return None;
-    }
-    let tmux_session_name = tmux_session_name
-        .map(str::trim)
-        .filter(|value| !value.is_empty())?;
-    if !crate::services::tmux_diagnostics::tmux_session_has_live_pane(tmux_session_name) {
-        return None;
-    }
-    let binding =
-        crate::services::tui_prompt_dedupe::runtime_binding_for_tmux_session(tmux_session_name)?;
-    if binding.runtime_kind != RuntimeHandoffKind::ClaudeTui {
-        return None;
-    }
-    let session_id = binding
-        .session_id
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())?;
-    if !std::path::Path::new(&binding.output_path).exists() {
-        return None;
-    }
-    Some(LiveTuiProviderSessionRecovery {
-        session_id: session_id.to_string(),
-        output_path: binding.output_path,
-    })
-}
-
-#[cfg(unix)]
-pub(super) async fn restore_live_tui_provider_session_from_binding(
-    shared: &Arc<SharedData>,
-    channel_id: serenity::ChannelId,
-    provider: &ProviderKind,
-    tmux_session_name: Option<&str>,
-    adk_session_key: Option<&str>,
-) -> Option<(String, bool)> {
-    let recovery = live_tui_provider_session_recovery(provider, tmux_session_name)?;
-    let memento_context_loaded = {
-        let mut data = shared.core.lock().await;
-        if let Some(session) = data.sessions.get_mut(&channel_id) {
-            session.restore_provider_session(Some(recovery.session_id.clone()));
-            session.memento_context_loaded
-        } else {
-            false
-        }
-    };
-    if let Some(session_key) = adk_session_key {
-        super::super::super::adk_session::save_provider_session_id(
-            session_key,
-            &recovery.session_id,
-            Some(&recovery.session_id),
-            provider,
-            channel_id,
-            shared.api_port,
-        )
-        .await;
-    }
-    let ts = chrono::Local::now().format("%H:%M:%S");
-    tracing::info!(
-        "  [{ts}] ↻ Recovered provider session_id from live TUI runtime binding for channel {}: tmux={} transcript={}",
-        channel_id.get(),
-        tmux_session_name.unwrap_or("(none)"),
-        recovery.output_path
-    );
-    Some((recovery.session_id, memento_context_loaded))
-}
-
-#[cfg(not(unix))]
-pub(super) async fn restore_live_tui_provider_session_from_binding(
-    _shared: &Arc<SharedData>,
-    _channel_id: serenity::ChannelId,
-    _provider: &ProviderKind,
-    _tmux_session_name: Option<&str>,
-    _adk_session_key: Option<&str>,
-) -> Option<(String, bool)> {
-    None
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(super) enum RuntimeMismatchVerdict {
-    Match,
-    Recreate,
-    Defer,
-}
-
-impl RuntimeMismatchVerdict {
-    pub(super) fn should_defer(self) -> bool {
-        self == Self::Defer
-    }
-}
-
-pub(super) fn runtime_mismatch_verdict(
-    expected: ManagedRuntimeExpectation,
-    observed: ObservedManagedRuntimeKind,
-    live_turn: bool,
-) -> RuntimeMismatchVerdict {
-    if expected.runtime_kind == observed.runtime_kind {
-        RuntimeMismatchVerdict::Match
-    } else if expected.evidence_strength == RuntimeKindEvidenceStrength::Weak || live_turn {
-        RuntimeMismatchVerdict::Defer
-    } else {
-        RuntimeMismatchVerdict::Recreate
-    }
-}
-
-#[cfg(unix)]
-fn managed_runtime_transcript_busy(
-    provider: &ProviderKind,
-    current_path: Option<&str>,
-    session_id: Option<&str>,
-    tmux_session_name: Option<&str>,
-) -> bool {
-    match provider {
-        ProviderKind::Claude => observe_claude_tui_transcript_state_for_session(
-            current_path,
-            session_id,
-            tmux_session_name,
-        )
-        .is_busy(),
-        ProviderKind::Codex => {
-            observe_codex_tui_rollout_state_for_cwd(current_path, tmux_session_name, session_id)
-                .is_busy()
-        }
-        _ => false,
-    }
-}
-
-#[cfg(unix)]
-#[derive(Debug)]
-struct RuntimeMismatchDeferState {
-    count: u32,
-    first_seen: std::time::Instant,
-    escalated: bool,
-}
-
-#[cfg(unix)]
-static RUNTIME_MISMATCH_DEFERS: std::sync::LazyLock<
-    dashmap::DashMap<(String, u64), RuntimeMismatchDeferState>,
-> = std::sync::LazyLock::new(dashmap::DashMap::new);
-#[cfg(unix)]
-const RUNTIME_MISMATCH_DEFER_ESCALATION_COUNT: u32 = 3;
-
-#[cfg(unix)]
-fn clear_runtime_mismatch_defer(provider: &ProviderKind, channel_id: serenity::ChannelId) {
-    RUNTIME_MISMATCH_DEFERS.remove(&(provider.as_str().to_string(), channel_id.get()));
-}
-
-#[cfg(unix)]
-fn record_runtime_mismatch_defer(
-    provider: &ProviderKind,
-    channel_id: serenity::ChannelId,
-    expected: ManagedRuntimeExpectation,
-    observed: ObservedManagedRuntimeKind,
-    open_inflight: bool,
-    transcript_busy: Option<bool>,
-    defer_reason: &'static str,
-) {
-    let key = (provider.as_str().to_string(), channel_id.get());
-    let mut state =
-        RUNTIME_MISMATCH_DEFERS
-            .entry(key)
-            .or_insert_with(|| RuntimeMismatchDeferState {
-                count: 0,
-                first_seen: std::time::Instant::now(),
-                escalated: false,
-            });
-    state.count = state.count.saturating_add(1);
-    let elapsed_secs = state.first_seen.elapsed().as_secs();
-    let escalating = !state.escalated && state.count >= RUNTIME_MISMATCH_DEFER_ESCALATION_COUNT;
-    if escalating {
-        state.escalated = true;
-    }
-    let count = state.count;
-    drop(state);
-
-    crate::services::observability::emit_inflight_lifecycle_event(
-        provider.as_str(),
-        channel_id.get(),
-        None,
-        None,
-        None,
-        if escalating {
-            "runtime_kind_mismatch_defer_escalated"
-        } else {
-            "runtime_kind_mismatch_deferred"
-        },
-        serde_json::json!({
-            "consecutive_defer_count": count,
-            "elapsed_secs": elapsed_secs,
-            "escalation_threshold": RUNTIME_MISMATCH_DEFER_ESCALATION_COUNT,
-            "action": "continue_defer_without_kill",
-            "defer_reason": defer_reason,
-            "expected_runtime_kind": expected.runtime_kind.as_str(),
-            "observed_runtime_kind": observed.runtime_kind.as_str(),
-            "open_inflight": open_inflight,
-            "transcript_busy": transcript_busy,
-        }),
-    );
-    if count == 1 || escalating {
-        tracing::warn!(
-            provider = provider.as_str(),
-            channel_id = channel_id.get(),
-            consecutive_defer_count = count,
-            elapsed_secs,
-            escalating,
-            "managed tmux runtime mismatch deferred without killing the session"
-        );
-    }
-}
-
-#[cfg(unix)]
-fn reconcile_managed_tmux_runtime_kind_for_config(
-    provider: &ProviderKind,
-    channel_id: serenity::ChannelId,
-    tmux_session_name: Option<&str>,
-    expected: Option<ManagedRuntimeExpectation>,
-    live_pane: impl FnOnce(&str) -> bool,
-    observe: impl FnOnce(&ProviderKind, &str) -> Option<ObservedManagedRuntimeKind>,
-    open_inflight: impl FnOnce() -> bool,
-    transcript_busy: impl FnOnce() -> bool,
-    recreate: impl FnOnce(&str, ManagedRuntimeExpectation, ObservedManagedRuntimeKind),
-) -> RuntimeMismatchVerdict {
-    let (Some(tmux_session_name), Some(expected)) = (tmux_session_name, expected) else {
-        clear_runtime_mismatch_defer(provider, channel_id);
-        return RuntimeMismatchVerdict::Match;
-    };
-    if !provider.uses_managed_tmux_backend() || !live_pane(tmux_session_name) {
-        clear_runtime_mismatch_defer(provider, channel_id);
-        return RuntimeMismatchVerdict::Match;
-    }
-    let Some(observed) = observe(provider, tmux_session_name) else {
-        clear_runtime_mismatch_defer(provider, channel_id);
-        return RuntimeMismatchVerdict::Match;
-    };
-    if observed.runtime_kind == expected.runtime_kind {
-        clear_runtime_mismatch_defer(provider, channel_id);
-        return RuntimeMismatchVerdict::Match;
-    }
-
-    let open_inflight = open_inflight();
-    let transcript_busy = (!open_inflight).then(transcript_busy);
-    let verdict = runtime_mismatch_verdict(
-        expected,
-        observed,
-        open_inflight || transcript_busy.unwrap_or(false),
-    );
-    if verdict.should_defer() {
-        record_runtime_mismatch_defer(
-            provider,
-            channel_id,
-            expected,
-            observed,
-            open_inflight,
-            transcript_busy,
-            if open_inflight {
-                "live_turn"
-            } else {
-                "weak_expected_evidence"
-            },
-        );
-        return verdict;
-    }
-    clear_runtime_mismatch_defer(provider, channel_id);
-    recreate(tmux_session_name, expected, observed);
-    verdict
-}
-
-#[cfg(unix)]
-pub(super) fn reconcile_managed_tmux_runtime_kind_using_runtime(
-    provider: &ProviderKind,
-    channel_id: serenity::ChannelId,
-    tmux_session_name: Option<&str>,
-    expected: Option<ManagedRuntimeExpectation>,
-    current_path: Option<&str>,
-    session_id: Option<&str>,
-) -> RuntimeMismatchVerdict {
-    reconcile_managed_tmux_runtime_kind_for_config(
-        provider,
-        channel_id,
-        tmux_session_name,
-        expected,
-        crate::services::tmux_diagnostics::tmux_session_has_live_pane,
-        observed_runtime_kind_for_managed_tmux,
-        || {
-            super::super::super::inflight::load_inflight_state(provider, channel_id.get())
-                .is_some_and(|state| !state.terminal_delivery_completed())
-        },
-        || managed_runtime_transcript_busy(provider, current_path, session_id, tmux_session_name),
-        |tmux_session_name, expected, observed| {
-            let reason = format!(
-                "tui_hosting config changed: expected {}, found {}; recreating tmux session",
-                expected.runtime_kind.as_str(),
-                observed.runtime_kind.as_str()
-            );
-            tracing::warn!(
-                provider = provider.as_str(),
-                channel_id = channel_id.get(),
-                tmux_session_name,
-                expected_runtime_kind = expected.runtime_kind.as_str(),
-                observed_runtime_kind = observed.runtime_kind.as_str(),
-                "managed tmux runtime kind mismatch detected; killing stale session before dispatch"
-            );
-            crate::services::termination_audit::record_termination_for_tmux(
-                tmux_session_name,
-                None,
-                "discord_dispatch",
-                "runtime_kind_mismatch_recreate",
-                Some(&reason),
-                None,
-            );
-            crate::services::tmux_diagnostics::record_tmux_exit_reason(tmux_session_name, &reason);
-            crate::services::platform::tmux::kill_session(tmux_session_name, &reason);
-            crate::services::tmux_common::cleanup_session_temp_files(tmux_session_name);
-            let cleared_runtime_binding =
-                crate::services::tui_prompt_dedupe::clear_tmux_runtime_binding(tmux_session_name);
-            tracing::debug!(
-                provider = provider.as_str(),
-                channel_id = channel_id.get(),
-                tmux_session_name,
-                cleared_runtime_binding,
-                "cleared stale tmux runtime binding after runtime kind mismatch"
-            );
-        },
-    )
-}
-
 pub(super) fn apply_prelaunch_runtime_kind(
     state: &mut InflightTurnState,
     runtime_kind: Option<RuntimeHandoffKind>,
@@ -918,216 +403,6 @@ pub(super) async fn reset_provider_session_after_worktree_isolation(
 #[cfg(test)]
 mod thread_role_inheritance_tests {
     use super::*;
-
-    #[test]
-    fn matrix_preserves_live_turns_and_only_recreates_strong_idle_mismatches() {
-        let cases = [
-            (
-                RuntimeKindEvidenceStrength::Weak,
-                RuntimeKindEvidenceStrength::Strong,
-                true,
-                RuntimeMismatchVerdict::Defer,
-                "weak expected plus strong observed live turn",
-            ),
-            (
-                RuntimeKindEvidenceStrength::Weak,
-                RuntimeKindEvidenceStrength::Weak,
-                false,
-                RuntimeMismatchVerdict::Defer,
-                "weak expected plus weak observed evidence",
-            ),
-            (
-                RuntimeKindEvidenceStrength::Strong,
-                RuntimeKindEvidenceStrength::Strong,
-                false,
-                RuntimeMismatchVerdict::Recreate,
-                "strong config change with no live turn",
-            ),
-            (
-                RuntimeKindEvidenceStrength::Strong,
-                RuntimeKindEvidenceStrength::Strong,
-                true,
-                RuntimeMismatchVerdict::Defer,
-                "strong config change with live turn",
-            ),
-            (
-                RuntimeKindEvidenceStrength::Weak,
-                RuntimeKindEvidenceStrength::Strong,
-                false,
-                RuntimeMismatchVerdict::Defer,
-                "strong observed evidence cannot authorize dispatch under a weak expectation",
-            ),
-            (
-                RuntimeKindEvidenceStrength::Strong,
-                RuntimeKindEvidenceStrength::Moderate,
-                false,
-                RuntimeMismatchVerdict::Recreate,
-                "moderate observed evidence preserves strong idle config transition",
-            ),
-        ];
-        for (expected_strength, observed_strength, live_turn, want, label) in cases {
-            let expected = ManagedRuntimeExpectation {
-                runtime_kind: RuntimeHandoffKind::LegacyTmuxWrapper,
-                evidence_strength: expected_strength,
-            };
-            let observed = ObservedManagedRuntimeKind {
-                runtime_kind: RuntimeHandoffKind::ClaudeTui,
-                evidence_strength: observed_strength,
-            };
-            assert_eq!(
-                runtime_mismatch_verdict(expected, observed, live_turn),
-                want,
-                "{label}"
-            );
-        }
-    }
-
-    #[test]
-    fn matching_runtime_never_recreates_regardless_of_liveness() {
-        for live_turn in [false, true] {
-            assert_eq!(
-                runtime_mismatch_verdict(
-                    ManagedRuntimeExpectation {
-                        runtime_kind: RuntimeHandoffKind::ClaudeTui,
-                        evidence_strength: RuntimeKindEvidenceStrength::Strong,
-                    },
-                    ObservedManagedRuntimeKind {
-                        runtime_kind: RuntimeHandoffKind::ClaudeTui,
-                        evidence_strength: RuntimeKindEvidenceStrength::Strong,
-                    },
-                    live_turn,
-                ),
-                RuntimeMismatchVerdict::Match
-            );
-        }
-    }
-
-    #[cfg(unix)]
-    fn reconcile_with_observation(
-        expected: ManagedRuntimeExpectation,
-        observed: ObservedManagedRuntimeKind,
-        live_turn: bool,
-        sink: &mut Vec<String>,
-    ) -> RuntimeMismatchVerdict {
-        reconcile_managed_tmux_runtime_kind_for_config(
-            &ProviderKind::Claude,
-            ChannelId::new(50_150_001),
-            Some("AgentDesk-5015-runtime"),
-            Some(expected),
-            |_| true,
-            |_, _| Some(observed),
-            || live_turn,
-            || false,
-            |name, _, _| sink.push(name.to_string()),
-        )
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn reconcile_match_does_not_call_recreate_sink() {
-        let runtime = ManagedRuntimeExpectation {
-            runtime_kind: RuntimeHandoffKind::ClaudeTui,
-            evidence_strength: RuntimeKindEvidenceStrength::Strong,
-        };
-        let mut calls = Vec::new();
-        assert_eq!(
-            reconcile_with_observation(
-                runtime,
-                ObservedManagedRuntimeKind {
-                    runtime_kind: RuntimeHandoffKind::ClaudeTui,
-                    evidence_strength: RuntimeKindEvidenceStrength::Strong,
-                },
-                false,
-                &mut calls,
-            ),
-            RuntimeMismatchVerdict::Match
-        );
-        assert!(calls.is_empty());
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn reconcile_defer_does_not_call_recreate_sink() {
-        let mut calls = Vec::new();
-        assert_eq!(
-            reconcile_with_observation(
-                ManagedRuntimeExpectation {
-                    runtime_kind: RuntimeHandoffKind::LegacyTmuxWrapper,
-                    evidence_strength: RuntimeKindEvidenceStrength::Strong,
-                },
-                ObservedManagedRuntimeKind {
-                    runtime_kind: RuntimeHandoffKind::ClaudeTui,
-                    evidence_strength: RuntimeKindEvidenceStrength::Strong,
-                },
-                true,
-                &mut calls,
-            ),
-            RuntimeMismatchVerdict::Defer
-        );
-        assert!(
-            calls.is_empty(),
-            "Defer must return before destructive wiring"
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn reconcile_recreate_calls_sink_once_for_exact_tmux_session() {
-        let mut calls = Vec::new();
-        assert_eq!(
-            reconcile_with_observation(
-                ManagedRuntimeExpectation {
-                    runtime_kind: RuntimeHandoffKind::LegacyTmuxWrapper,
-                    evidence_strength: RuntimeKindEvidenceStrength::Strong,
-                },
-                ObservedManagedRuntimeKind {
-                    runtime_kind: RuntimeHandoffKind::ClaudeTui,
-                    evidence_strength: RuntimeKindEvidenceStrength::Strong,
-                },
-                false,
-                &mut calls,
-            ),
-            RuntimeMismatchVerdict::Recreate
-        );
-        assert_eq!(calls, ["AgentDesk-5015-runtime"]);
-    }
-
-    #[test]
-    fn prelaunch_claude_tui_seed_omits_wrapper_output_but_preserves_fifo() {
-        let seed = prelaunch_inflight_runtime_seed_from_paths(
-            "AgentDesk-claude-seed",
-            "/runtime/wrapper-stream.log".to_string(),
-            "/runtime/input.fifo".to_string(),
-            true,
-            Some(RuntimeHandoffKind::ClaudeTui),
-        );
-        assert_eq!(seed.0.as_deref(), Some("AgentDesk-claude-seed"));
-        assert_eq!(
-            seed.1, None,
-            "ClaudeTui must wait for RuntimeReady transcript binding"
-        );
-        assert_eq!(seed.2.as_deref(), Some("/runtime/input.fifo"));
-        assert_eq!(seed.3, 0);
-    }
-
-    #[test]
-    fn prelaunch_seed_is_identical_for_intake_and_headless_callers() {
-        let intake = prelaunch_inflight_runtime_seed_from_paths(
-            "AgentDesk-claude-symmetric",
-            "/runtime/wrapper-stream.log".to_string(),
-            "/runtime/input.fifo".to_string(),
-            true,
-            Some(RuntimeHandoffKind::ClaudeTui),
-        );
-        let headless = prelaunch_inflight_runtime_seed_from_paths(
-            "AgentDesk-claude-symmetric",
-            "/runtime/wrapper-stream.log".to_string(),
-            "/runtime/input.fifo".to_string(),
-            true,
-            Some(RuntimeHandoffKind::ClaudeTui),
-        );
-        assert_eq!(intake, headless);
-    }
 
     #[test]
     fn managed_linked_worktree_skips_provider_reisolation_without_session_state() {
